+++ tests/Commands/WorkCommandTest.php
@@ -7,6 +7,7 @@ namespace AvtoDev\AmqpRabbitLaravelQueue\Tests\Commands;
 use AvtoDev\AmqpRabbitLaravelQueue\Worker;
 use AvtoDev\AmqpRabbitLaravelQueue\Commands\WorkCommand;
 use AvtoDev\AmqpRabbitLaravelQueue\Tests\AbstractTestCase;
+use Symfony\Component\Console\Input\InputOption;
 
 /**
  * @covers \AvtoDev\AmqpRabbitLaravelQueue\Commands\WorkCommand<extended>
@@ -29,5 +30,22 @@ class WorkCommandTest extends AbstractTestCase
 
             $this->assertInstanceOf(Worker::class, $worker);
         }
+    }
+
+    /**
+     * @small
+     *
+     * @return void
+     */
+    public function testCustomizedCommandOptions(): void
+    {
+        /** @var WorkCommand $command */
+        $command = $this->app->make('command.queue.work');
+
+        /** @var InputOption[] $options */
+        $options = $command->getDefinition()->getOptions();
+
+        $this->assertRegExp('~not used~i', $options['sleep']->getDescription());
+        $this->assertEquals(-1, $options['timeout']->getDefault());
     }
 }