+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 ### Changed
 
 - `Queue` class constructor signature
+- Option `sleep` for `queue:work` command marked as unused
+- Option `timeout` for `queue:work` now `-1` by default. It means next - by default used timeout value from configuration file, but this value can be overridden by passing `--timeout` option with `0..+n` value
 
 ## v1.0.1